requirements.txt
CacheControl==0.12.11
cachetools==5.2.0
cachy==0.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
cryptography==3.4.8
distlib==0.3.5
filelock==3.7.1
google-auth==2.11.0
google-auth-oauthlib==0.5.2
greenlet==1.1.3
gspread==5.5.0
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
jeepney==0.8.0
keyring==23.8.2
lockfile==0.12.2
msgpack==1.0.4
numpy==1.23.2
oauthlib==3.2.0
packaging==20.9
pandas==1.4.4
pastel==0.2.1
pexpect==4.8.0
pipenv==2022.8.5
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.1.14
poetry-core==1.0.8
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pylev==1.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
SecretStorage==3.3.2
shellingham==1.5.0
six==1.16.0
SQLAlchemy==1.4.41
tomlkit==0.11.2
urllib3==1.26.11
virtualenv==20.16.3
virtualenv-clone==0.5.7
webencodings==0.5.1
xmltodict==0.13.0
zipp==3.8.1
